chemistry/requirements.txt
asgiref==3.8.1
Django==5.1.5
gunicorn==23.0.0
packaging==24.1
sqlparse==0.5.1
tzdata==2024.2
qrcode==8.0
Pillow==11.1.0
pytz==2025.1
mysqlclient
django-autocomplete-light
django-filter